requirements.txt
flask
Flask-Session
python-dotenv
requests
sounddevice
scipy
psutil
openai
scikit-learn
tqdm
numpy
markdown
pdfplumber
faiss-cpu
cryptography
jsonschema
# Wake word + VAD
openwakeword
webrtcvad